requirements.txt
absl-py==0.9.0
Ale==0.2.0
astor==0.8.1
cachetools==4.0.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
docker==4.2.0
future==0.18.2
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.1
gym-retro==0.7.0
h5py==2.10.0
idna==2.8
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.3.2
pypiwin32==219
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tqdm==4.42.1
urllib3==1.25.8
websocket-client==0.57.0
Werkzeug==1.0.0
wincertstore==0.2
wrapt==1.11.2
